added new interface function to wrapper_out to add meta-data about data sets
moved initialize and close wrapper out of hash_case functions,
it now expects a wrapper that has been initialized.  This was
changed to make it simpler to add meta data about data sets.

This changes the spec of the data files, wrapper_in will need be changed
to access the meta data eventually, and wrapper_in_hdf needs to be changed to deal
with the new naming scheme

git-svn-id: svn://innoue/tracking/trunk@158 4cb6fdf0-3dc9-4801-8ffc-c6659000aa44