Fixed issue with FreeImage seg faulting.
The issue was some bad interaction between fipImage and fipMultipage

The former unloads FIBITMAP objects with some of it's function, but the
latter requires that you do not unload the FIBITMAP objects it hands out.

The best solution is to fix the FIP classes, but that is long term.
To get this working I am using the C interface for the bitmaps instead
of the c++ interface.  This is only really an issue if you feed in something that
is not uint16 and need to gray scale it